pyproject.toml
[project]
name = "redbot"
version = "0.1.0"
description = ""
authors = [
    {name = "RedR1ghtHand",email = "4aikaman@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest (>=6.2.3,<7.0.0)",
    "discord-py (>=2.6.3,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pymongo (>=4.15.3,<5.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pydantic (>=2.12.2,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
